fix unsigned variable with -1 as flag value
git-svn-id: svn+ssh://svn.code.sf.net/p/openlcb/svn/trunk/prototypes/C@2711 87536cb5-86f1-4cdc-8fc6-3f0f71bcd267

+++ OlcbDesktopCAN/BG.h
@@ -48,7 +48,7 @@ class BG {
   ButtonLed* blue;
   ButtonLed* gold;
 
-  uint8_t index;
+  int16_t index;
   bool lastBlue;
   bool lastGold;
   bool started;
